requirements.txt
mss
Pillow
imageio
imageio-ffmpeg
